requirements.txt
flask==1.0.2
flask-restful==0.3.7
flask_cors
python-dotenv
gunicorn
responses
pytest
mock